wish: SourceForge compatibility
Last week I noticed that SourceForge runs some variant of Mailman for
its mailing lists, so I tried to run my WWW::Mailman-based management
script on one of its list.

It does not work.  One stumbling block is the hardcoded requirement for
'mailman' to be in the URL for the Mailman installation, but working
around that (see the patch) wasn't enough, so I gave up.

It would be nice if WWW::Mailman was Sourceforge-compatible.

URL: https://rt.cpan.org/Ticket/Display.html?id=73312