a first stab at the admindb() method
This only returns the form itself, which is very complex and contains
a lot of data. We should return extra information either by parsing
the whole page, or maybe returning a list of links to individiual
emails (with the msgid query-string parameter).

It should also be noted that the form changed in Mailman's lifetime,
so we might need to change behaviour depending on the version number
(depending how far in the past we want to support Mailman).